Large code rewrite for cleanliness.  No semantic change.
Beauty is in the eye of the beholder, so Caveat Emptor.

Broke up the RadixOrder file into 4 files; RadixOrder, RadixCount,
SingleThreadRadixOrder, and SplitByMSBLocal.

Broke out most uses of _left and _right in BinaryMerge into one of two
structs; a left/right struct available at constructor time (hence is
shared amongst all nodes and shallow-copied around the cluster), and a
left/right struct made in each compute2() call.  Then moved lots of
duplicated code into one of those 2 structs (and cut the replicated
half out).

Added a at8order accessor to remove a common mod/div 2D array lookup.
Flowed long comments, aligned code.  Removed ideaJ warnings.  Made
much code final, private or both.
